+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `tableColumns` function
  - function aliases
  - `deprecated` wrapper
+ - `promiseChain` function
 
 ### Changed
  - use `ts-jest` instead of `jest`